refactor: 헤더, 푸터 디렉토리 생성
