+++ src/app/menu/page.tsx
@@ -5,7 +5,7 @@ import MenuCategoryList from '@/components/pages/menu/MenuCategoryList';
 import { categoryMenus } from '@/data/dummyDatas';
 import MenuTop from '@/components/pages/menu/MenuTop';
 import MenuBannerList from '@/components/pages/menu/MenuBannerList';
-import Header from '@/components/layouts/Header';
+import Header from '@/components/layouts/header/Header';
 
 export default function MenuPage() {
   return (
